refactor(domain): update entity props imports to use shared types
