+++ src/domain/entities/question/ports/question.props.ts
@@ -1,4 +1,4 @@
-import type { Optional } from '@/util/types/optional'
+import type { Optional } from '@/shared/types/common/optional'
 import type { Question } from '../question.entity'
 
 export type QuestionProps = Optional<Omit<Question, 'id' | 'slug' | 'answers'>, 'createdAt' | 'updatedAt' | 'bestAnswerId'>
