+++ src/domain/entities/refresh-token/ports/refresh-token.props.ts
@@ -1,4 +1,4 @@
-import type { Optional } from '@/util/types/optional'
+import type { Optional } from '@/shared/types/common/optional'
 import type { RefreshToken } from '../refresh-token.entity'
 
 export type RefreshTokenProps = Optional<Omit<RefreshToken, 'id'>, 'createdAt' | 'updatedAt' | 'expiresAt'>
